Unified all rosbag apps
